move slightfile to its own crate
