+++ crates/core/Cargo.toml
@@ -11,12 +11,12 @@ doctest = false
 
 [dependencies]
 anyhow = { workspace = true }
-serde = { workspace = true }
 short-crypt = "1"
 rand = { workspace = true }
 toml = { workspace = true }
 clap = { workspace = true }
 semver = { workspace = true }
+slight-file = { workspace = true }
 
 [dev-dependencies]
 tempdir = { workspace = true }
