+++ crates/core/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod interface_parser;
 pub mod secret;
-pub mod slightfile;
 pub mod wasm_parser;
