+++ crates/core/src/secret.rs
@@ -4,10 +4,10 @@ use std::{
     io::Write,
 };
 
-use crate::slightfile::{Config, TomlFile};
 use anyhow::{bail, Result};
 use rand::{distributions::Alphanumeric, thread_rng, Rng};
 use short_crypt::ShortCrypt;
+use slight_file::{Config, TomlFile};
 
 pub const SLIGHTKEY: &str = ".slightkey";
