--- crates/core/src/slightfile.rs
@@ -1,30 +1,0 @@
-use std::collections::HashMap;
-
-use serde::{Deserialize, Serialize};
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct TomlFile {
-    pub specversion: String,
-    pub secret_store: Option<String>,
-    pub secret_settings: Option<Vec<Config>>,
-    pub capability: Option<Vec<Capability>>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Capability {
-    pub resource: Option<String>,
-    pub name: String,
-    pub configs: Option<HashMap<String, String>>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Config {
-    pub name: String,
-    pub value: String,
-}
-
-impl Config {
-    pub fn new(name: String, value: String) -> Self {
-        Self { name, value }
-    }
-}
