+++ crates/slightfile/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "slight-file"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow = { workspace = true }
+serde = { workspace = true }
